Started working on Abilities that trigger on contact
By adding a few move arrays with list of spore, punching and contact moves, and
moving CheckIfXIsXType to home/battle.asm to not needlessy duplicate code. This can
also be used in the AI instead of duplicating type checks.

Reversing the change would be easy enough if home/battle.asm would get full.