Fixed bugged type matchups
When Compound Eyes and Synchronize was given effects against wild Pokémon, it used GetAbility.
GetAbility overrides CurSpecies (it does this to get the ability info). This lead to part of
the Pokémon data structure for wild pokémon getting its data based on the 1st Pokémon in your
party. This was solved by making GetAbility restore whatever was there beforehand when done.