Refactored type matchup handling
The actual type matchup function wasn't altered much, but the way the game
used the function in the actual damage calculation was made far less awkward,
and this removed a lot of overcomplication and duplication. Hopefully, there
weren't any residual bugs, but even if there was, this should be much easier
to work with.