Fix part of #256 - cap recovery for drain moves
Drain moves base their recovery on the floor of 1/2 of wCurDamage (or 1, if only 1 damage was dealt). In base Crystal, there is code in the subtract HP function that sets wCurDamage to the original value of wPlayerMonHP/wEnemyMonHP before changing those values if the Pokemon takes more damage than it has remaining HP (in other words, set wCurDamage to what the Pokemon's HP was before it fainted). This fix re-implements that code, which will also make sure Pokemon take correct damage on recoil damage (which also bases its recoil damage on wCurDamage). This adds 32 cycles to each damage calculation.