properly check if mon faints on switchout move
Ties in with the modifications to core.asm, if a pokemon faints after getting hit by U-Turn or Volt Switch, we need extra checks to determine who fainted and if a switch even still needs to happen

Basically the checking happens as follows:
Did the opponent faint?
Was it the player's pokemon or the opponents pokemon who fainted?
Call the relevant fainting handler (faint the opponent and give EXP if needed), then return to the switchout function before the opponent post-faint switch script occurs
Is the battle over? If so, no need to do any switching.
If it's not over, finish the switchout then finish the fainting handler, letting the user/NPC select their next pokemon after they see who switched in for their opponent.