Fix Roaming Pokémon encounter bug
If the encounter system declared that you were to encounter a roaming Pokémon,
but the encounter fails because the encounter rate is too low, the roaming
Pokémon would get initialized with bad data, and the roaming Pokémon would
leave the route without being encountered.

This bug was introduced when the encounter rate check was moved to happen after
checking for wild Pokémon. This change was made to make various overworld
abilities work, hence it can't just be undone.