Fix #421: After Kurt makes a Ball, healing in the Pokémon Center turns the player sprite black
Usually the sequence of Special_FadeBlackQuickly + Special_ReloadSpritesNoPalettes is followed by Special_FadeInQuickly. Here it is not, and all-black zeros were left in the RED and BLUE palettes of wOBPals2.

As far as I can tell, Special_ReloadSpritesNoPalettes does not need to be writing to wOBPals2 at all. This may be a latent bug in pokecrystal.