Optimize LCD HBlank code, fixing Music Player bugs
I didn't really look closely at the Music Player code, figuring that it had
plenty of time to do its thing considering that the music player is a thing
in other hacks that are single-speed. Well, apparently this assumption was
wrong (maybe it temporarily switches speed in those?) and the difference in
speed as a result of more checks during LCD H-Blank was enough to cause the
green notes in the music player to have its graphics screwed up. I tweaked
the code so that it would use the existing hram jump instruction to be able
to skip the hMPState checks entirely. As a result, the music player and the
other hblank codepaths now has slightly more cycles to work with.