Fix stat calculation in BT.
For whatever reason, it takes EVs - 1 as parameter, not EVs, which led to it
reading the incorrect EVs. This also led to enemy HPs being drastically lower
than they should in some cases (since it was reading that for its HP EVs), but
this was less noticeable unless you paid attention. The same also applied to
your HP of course, but since not all players ran with 252 HP EVs there in first
place...