Add Alolan Vulpix+Ninetales to Seafoam Islands.
These Pokémon are more active in snowy mountain ranges. Unfortunately,
this isn't Sinnoh, so we place them in Seafoam instead. This way,
we have Sandshrew-A and its evolution in Ice Path and Vulpix-A and its
evolution in Seafoam Islands. I did restrict Vulpix+Ninetales from the
lower Seafoam floors, so that they're closer to the surface.