Implement the newbox PC storage system (#608)
(Squashed from the `newbox` branch)

Many thanks to FIQ for rewriting the backend to (a) navigate all the boxes without saving, and (b) store 16 boxes instead of 14 (320 Pokémon instead of 280)! Also for a Gen 3+-style frontend with colored icons and Menu/Swap/Item modes. And to darsh for writing BSP scripts so that Prism's `bspcomp` can patch old save files as far back as 2019.

The `box_struct` is no longer a subset of `party_struct`; instead, `party_struct` is encoded to `savemon_struct` on deposit and decoded on withdraw. (They still begin with mostly the same fields, but `savemon_struct` stores a single PP Up byte and has 7-bit encoded nickname and OT fields with a checksum in the 8th bits. This means that PP as well as HP and status will be restored on deposit.)

Long-term plans include using redrawn icons by choosh based on Crystal Clear, with the same palettes as full-size sprites; and rearranging more of the struct fields to allow 10-bit mon, move, and item IDs (up to 510 of each), 20-bit OT ID (6 digits, 000000-999999), or even 5-bit IVs (0-31 instead of 0-15).

Fixes #510

Co-authored-by: Fredrik Ljungdahl <fredde1994@gmail.com>
Co-authored-by: itsdarsh <twitteruserdarsh@gmail.com>