Restore the opponent party for research data if user aborts a party select.
If the player selects 3 rental Pokémon to enter, legality checking is performed
by storing the selected party in the OTParty table. If, after this, the player
changes their mind and attempts to abort party selection, but returns to it,
we need to restore the opponent trainer data to the party again so we can give
accurate research data.