Fix Cerulean Cave becoming inaccessible
Before this fix, using an escape rope (or Dig) while crossing the bridge in Cerulean Cave 1F will not reset the scene ID that determines whether the player is passing over or under the bridge. There also are no coord events to handle this on the water underneath the bridge. Therefore, if a player escapes then tries to surf back, the bridge will still be in underfoot mode and thus cannot be crossed, rendering the rest of Cerulean Cave inaccessible without hacks. Moving the Overhead coord events to the water next to the bridge resolves this problem.

Also, I moved the coord even to turn the bridge into underfoot mode one tile down in the Y-direction. This fixed a minor graphical bug where the player walks up to the bridge and is still "underfoot" for that first step, being partially drawn under the bridge.