Implement anim_setbgpal/anim_setobjpal in the battle animation system
These commands can be used to set a particular BG/OBJ pal to a custom
palette that can be defined for theoretically up to 256 palettes.
The first parameter denotes the BG/OBJ pal index, the 2nd the battle pal to use.
No palettes are actually defined (beyond a grayscale test one) in this commit.
If the first parameter is ANIM_BATTLE_USER or ANIM_BATTLE_TARGET, all palettes
used for the corresponding user or target is set.