fix XY title music distortion
Channel 1 needs to start with a rest note to clear a wavelength sweep, so the original title track works fine. Not so for the XY track. Unfortunately the options are either to add a rest note in front of every channel for the XY theme to maintain proper timing, or do something dumb and hacky. Guess which one I chose.