Consistency Tweaks to parties.asm (#883)
* Update evos_attacks.asm for proper Eevee location evolutions

In line with other fixes to the order of evolutions listed, this change lists the location evolutions for Glaceon and Leafeon before (instead of after) the happiness evolution conditions for Espeon and Umbreon. As the code stands, a sufficiently happy Eevee will not evolve into Glaceon or Leafeon, and rather attempt to become Espeon or Umbreaon, even if the player is in the correct location.

* Consistency Tweaks to parties.asm 

Changed Machamp's moveset in the first Bruno fight from having Swords Dance to Bulk Up as SD is not in Machamp's learnset pool.

Also swapped the Kadabra and Xatu for Psychic Phil based on the  pinned comment in #Testing-9Bit. on discord.