Goldenrod Museum fix
Fixes Goldenrod Museum pictures crashing the games. I've used the dba macro instead of fardw in the pic list. I'm not knowledgeable enough to understand why the fardw macro was having a hard issue loading the pic correctly, as that's what's leading to crashes. It only saves 7 bytes, so I'm not sure it's worth it.