put footprint queue in dedicated memory in wram0
The issue described in #903 was the canary in the coalmine for a serious issue, where every VBlank0 `PlaceFootprints` had the potential to overwrite arbitrary sections of memory with garbage data if the first byte of the queue was nonzero. This could lead to save file corruption and potentially crashes. The quick and dirty fix is just to give the queue its own section in memory, though committing 21 tiles to this feels bad. An alternative may be to have a dedicated "footprints to draw" counter, which would allow the queue itself to be in shared memory rather harmlessly. Of course, that may not be trivial itself, and for now I just want to avoid potential game-breaking problems.