Orange Islands Overhaul
This aims to properly cement Orange Islands as the 'de facto' endgame area, by both boosting the badge-base-level curve to cap at 68 instead of 61, and by drastically boosting the level of the wild Pokemon in Orange Islands. Additionally this adds the badge system to every trainer there and adds parties to the TODO bulbasaur trainers.