Update evos_attacks.asm (#971)
Edited the following lines to the movesets in Polished Crystal to fix various errors, typos, and redundancies. No other changes have been made.

Psyduck: Amnesia was originally at Level 47 and was out of order with the rest of its moves, this is now at Level 37 and is in order. (Line 1208)

Primeape: A minor change, but Rage is now listed as an evolution move in the comments, which it is. (Line 1261)

Poliwhirl: Low Kick was at Level 28, which is the same place as Poliwag despite the moves before and after both being at a higher level after evolving. This was moved to Level 32 to be consistent with the base games in Gen 6 and the pattern of level ups created. (Line 1385)

Dodrio: Rage was listed at Level 9 instead of 8, which makes no sense as that's after it evolves. For consistency's sake, I moved it back anyways. (Line 2020)

Jolteon: Learns Double Kick twice. At level 9 from Let's Go, and at level 17 from its normal moves. Since this isn't quite fair compared to the other Eeveelutions, Spark was added for Level 17 at the recommendation of "censoredharp" on the Discord. Its more powerful than Bite, but paralyzes rather than flinches (which Jolteon can already do), and with the lower Attack stat for Jolteon. This seemed like a balanced choice to keep a consistent movepool without adding extra capabilities Jolteon didn't have before. (Line 3222)

Tangrowth: Moved Swords Dance from Level 49 to 48 to be consistent with the mainline games. (Line 6063)

There were other ones I had considered, like Kabuto learning Leech Life like Kabutops instead of Giga Drain (Line 4818), and Wyrdeer's last 3 moves being moved down to match Stantler's (Lines 6360-6363), but I decided not to as these seemed more intentional, and not exactly necessary compared to Jolteon's case.

Co-authored-by: vulcandth <vulcandth@gmail.com>