ineligible switch textboxes appear on first turn
Increments `wTotalBattleTurns` at start of turn instead of end of turn, making it a counter of the current turn rather than number of turns passed, in effect. Allows for "An Egg can't battle!" and "There's no will to battle!" textboxes to appear on the first turn, they'd be skipped otherwise. This does mean anything that scales with the number of turns passed in a battle will be capped at 254 turns instead of 255, but that doesn't affect anything presently in Polished.