orange islands overhaul
Easier to work on now, this aims to cement the Orange Islands region as the "de facto" endgame area, by redoing all the trainers and wild battles to use the badge level system, then boosting their levels to match what was originally set. This does increase the difficulty of some trainers by a tiny bit, but overall the difficulty curve is the same if you venture there immediately after unlocking the area, and will remain relevant to visit any time thereafter. Also raised the badge level curve to continue as a straight line rather than taper off towards the end.