Rename some ability labels referencing "animations"
When I first implemented abilities in Polished Crystal, I needed a way to
disable move animations with a flag so I could reuse battle commands for
various abilities (such as Drizzle sharing code with Rain Dance). Since then,
wAnimationsDisabled does more things such as keeping track of what slideouts
have been activated (to avoid them triggering multiple times) or caching the
interface when enabled so that slideouts can be easily dismissed. On top of
this, wAnimationsDisabled implied that it was a generic "disable battle anims"
flag, which was never true in the first place. This renaming addresses this,
alongside renaming DisableAnimations/EnableAnimations to the more appropriate
BeginAbility/EndAbility.