Convert Psychic Inver's party table to the new format
Because I intend to separate DV data into its own table with parties only
referencing the relevant DV entry, I had to move Inver's party to be included
as part of the regular party table. Because it and the rest of Inver handling
must be in the same ROM bank, I added the 2 relevant sections into layouts.link.