Reduce size of conversion table.
This makes the conversion table size $100 instead of $200. Later we may assess if we can make it smaller.. but note, the minimum value is $100, but the unused space will be in ___padding which may be carefully used for other user code. Currently the table is now sized for no padding. Reference: https://github.com/aaaaaa123456789/pokecrystal16/wiki/16-bit-conversion-table "Declaration and layout" for more information.