Fix Destiny Bond/Spikes/Toxic Spikes targeting
Destiny Bond is self-targeting and shouldn't check for a valid target at all.
(Toxic) Spikes do have a target for the purpose of Pressure, but should work
even when the Pokémon on the target side is fainted.

This also fixes an issue where Pressure could proc for a fainted mon.