Remove dummy NO_ITEM from Bugsy's Butterfree
This was originally added when I first created the macros in order to match
the output binary from pre-macros, which I had as a goal in order to ensure
that there wasn't any bugs present in the trainer data parsing. Since the
code has definitely changed since then, the dummy entry no longer serves a
purpose.