Allow specifying a different trainer pal (#1082)
This implementation has the drawback of an extra byte for every trainer... so we loose some space, but the benefit is it allows us to change the trainer's frontpic palette similarly to the overworld object palette. This allows us to have the overworld palette match the frontpic palette and add more color variety.